Use single combination of access mode
Using 'ReadWriteOncePod' with others access modes causing a conflict.
Basing on CDI usage to determine the access mode based on storage
profile, we should fetch only one combination, basing on the storage
profile is sorted to 'Block' and 'ReadWriteMany' by CDI when creating a
storage profile. If the user overrides the defaults, we will respect it.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>