MTV-1966 | Allow to skip shared disks
Issue:
In the current release of Forklift it is possible to import shared disks using cold migration.
This is possible because the virt-v2v is transferring all disks attached to the VM, but because
the shared disks are attached to multiple VMs the shared disks get migrated multiple times and users right now,
need to delete the extra volumes and reattach the disks.

Fix:
Add a new parameter `migrateSharedDisks` to the plan which will determine if the shared disks should be migrated.
This parameter will allow the users to choose if the plan should migrate shared disks or not.
When the `migrateSharedDisks` will be enabled the Forklift will use the "normal" cold migration flow using virt-v2v and labeling
the shared PVC.
When the `migrateSharedDisks` will be disabled the Forklift will use the KubeVirt CDI for disk transfer, and it will skip
the shared disks. After the disk transfer it will try to locate the already shared PVCs and attach them to the VMs.
The Forklift will try to automatically find migrated shared PVCs in the namespace and attach them to the VMs.

User flow:
- Turn off all VMs with the attached shared disk on the VMware side.
- Create a plan with a single VM and the `migrateSharedDisks` enabled in the Forklift.
- Start the migration of the first plan and wait for it to finish.
- Create a secondary plan with all other VMs attached and the `migrateSharedDisks` disabled to the same target namespace as first plan.
- Start the migration of the second plan and wait for it to finish.
- Check if all shared disks are attached to all VMs.

Known issues:
- The shared disks are not passed to the guest conversion so virt-v2v is
  not aware of them and does not update the configfiles, for example
  fstab. This will be fixed in followup PR.
- The shared disks do not support warm migration.
- The shared disks can not be migrated to multiple target namepsaces.

Ref:
- Original issue: https://issues.redhat.com/browse/MTV-1966
- This is continuation of: https://github.com/kubev2v/forklift/pull/1318

Signed-off-by: Martin Necas <mnecas@redhat.com>