feat: Introduce vsphere xcopy volume populator (#1369)
feat: Introducing vSphere generic copy-offload

This commit introduces a new volume populator,
`VsphereXcopyVolumePopulator`,
designed to leverage vSphere's XCOPY functionality for efficient volume
cloning
during VM migrations.

This enhancement significantly improves the performance of volume
cloning for
vSphere VMs, especially for large volumes, by utilizing the server-side
XCOPY
operation.

When the feature flag `feature_copy_offload` is true (off by default),
the controller will
consult a config map to decided if VM disk from VMWare could be copied
by the storage backend(offloaded) into the newly created PVC.
When the controller creates the PVC for the v2v pod it will also create
a volume popoulator resource of type VSphereXcopyVolumePopulator and set
the filed `dataSourceRef` in the PVC to reference it.

Added a new populator controller for the resource
VSPhereXcopyVolumePopulator

A new populator implementation under cmd/vsphere-xcopy-volume-populator
is a cli program that runs in a container that is responsible to perform
XCOPY to effciently copy data from a VMDK to the target PVC. See the
flow in cmd/vsphere-xcopy-volume-populator/README.md
The populator uses the storage API (configurable) to map the PVC to an
ESX
then uses Vsphere API to call functions on the ESX to perform the actual
XCOPY command (provided that VAAI and accelerations is enabled on that
ESX).

If a storage provider wants to their storage to be supported they need
to implement a go package named after their product, and mutate main so
their specific code path is initialized. See
cmd/vsphere-xcopy-volume-populator/internal/populator/storage.go

- currently a VM with single disk is supported

---------

Signed-off-by: Roy Golan <rgolan@redhat.com>
Signed-off-by: Martin Necas <mnecas@redhat.com>
Co-authored-by: borod108 <boris.od@gmail.com>
Co-authored-by: Martin Necas <mnecas@redhat.com>